+++ node/src/DataProducer.ts
@@ -4,8 +4,9 @@ import { Channel } from './Channel';
 import { PayloadChannel } from './PayloadChannel';
 import { TransportInternal } from './Transport';
 import { SctpStreamParameters } from './SctpParameters';
+import { AppData } from './types';
 
-export type DataProducerOptions =
+export type DataProducerOptions<DataProducerAppData extends AppData = AppData> =
 {
 	/**
 	 * DataProducer id (just for Router.pipeToRouter() method).
@@ -31,7 +32,7 @@ export type DataProducerOptions =
 	/**
 	 * Custom application data.
 	 */
-	appData?: Record<string, unknown>;
+	appData?: DataProducerAppData;
 };
 
 export type DataProducerStat =
@@ -76,7 +77,8 @@ type DataProducerData =
 
 const logger = new Logger('DataProducer');
 
-export class DataProducer extends EnhancedEventEmitter<DataProducerEvents>
+export class DataProducer<DataProducerAppData extends AppData = AppData>
+	extends EnhancedEventEmitter<DataProducerEvents>
 {
 	// Internal data.
 	readonly #internal: DataProducerInternal;
@@ -94,7 +96,7 @@ export class DataProducer extends EnhancedEventEmitter<DataProducerEvents>
 	#closed = false;
 
 	// Custom app data.
-	readonly #appData: Record<string, unknown>;
+	#appData: DataProducerAppData;
 
 	// Observer instance.
 	readonly #observer = new EnhancedEventEmitter<DataProducerObserverEvents>();
@@ -115,7 +117,7 @@ export class DataProducer extends EnhancedEventEmitter<DataProducerEvents>
 			data: DataProducerData;
 			channel: Channel;
 			payloadChannel: PayloadChannel;
-			appData?: Record<string, unknown>;
+			appData?: DataProducerAppData;
 		}
 	)
 	{
@@ -127,7 +129,7 @@ export class DataProducer extends EnhancedEventEmitter<DataProducerEvents>
 		this.#data = data;
 		this.#channel = channel;
 		this.#payloadChannel = payloadChannel;
-		this.#appData = appData || {};
+		this.#appData = appData || {} as DataProducerAppData;
 
 		this.handleWorkerNotifications();
 	}
@@ -183,17 +185,17 @@ export class DataProducer extends EnhancedEventEmitter<DataProducerEvents>
 	/**
 	 * App custom data.
 	 */
-	get appData(): Record<string, unknown>
+	get appData(): DataProducerAppData
 	{
 		return this.#appData;
 	}
 
 	/**
-	 * Invalid setter.
+	 * App custom data setter.
 	 */
-	set appData(appData: Record<string, unknown>) // eslint-disable-line no-unused-vars
+	set appData(appData: DataProducerAppData)
 	{
-		throw new Error('cannot override appData object');
+		this.#appData = appData;
 	}
 
 	/**