+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 * `RtpStreamSend`: Reset RTP retransmission buffer upon RTP sequence number reset ([PR #1041](https://github.com/versatica/mediasoup/pull/1041)).
 * `RtpRetransmissionBuffer`: Handle corner case in which received packet has lower seq than newest packet in the buffer but higher timestamp  ([PR #1044](https://github.com/versatica/mediasoup/pull/1044)).
-* Node: Make `appData` TS typed and writable ([PR #XXXX](https://github.com/versatica/mediasoup/pull/XXXX), credits to @mango-martin).
+* Node: Make `appData` TS typed and writable ([PR #1046](https://github.com/versatica/mediasoup/pull/1046), credits to @mango-martin).
 
 
 ### 3.11.19